+++ tests/settings.py
@@ -186,6 +186,14 @@ ALLOWED_HOSTS = ["*"]
 
 EMAIL_BACKEND = "django.core.mail.backends.console.EmailBackend"
 
+# Make sure the search index is updated immediately during tests
+TASKS = {
+    "default": {
+        "BACKEND": "django_tasks.backends.immediate.ImmediateBackend",
+        "ENQUEUE_ON_COMMIT": False,
+    }
+}
+
 
 with contextlib.suppress(ImportError):
     from .local import *  # noqa: F403